Refactor main UI script name.
