+++ game/main.py
@@ -4,7 +4,7 @@ from panda3d.core import (
     load_prc_file
 )
 
-from ui.neoit import NeoImpressiveTitleUI
+from ui.neoimpressivetitleui import NeoImpressiveTitleUI
 from game_state import GameState
